allow bigger test arguments

+++ src/test/java/de/tilman_neumann/jml/factor/CombinedFactorAlgorithm1ThreadTest.java
@@ -13,6 +13,7 @@
  */
 package de.tilman_neumann.jml.factor;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.junit.BeforeClass;
@@ -34,7 +35,7 @@ public class CombinedFactorAlgorithm1ThreadTest extends FactorTestBase {
 	
 	@Test
 	public void testSmallestComposites() {
-		List<Integer> fails = testFullFactorizationOfComposites(100000);
+		List<BigInteger> fails = testFullFactorizationOfComposites(100000);
 		assertEquals("Failed to factor n = " + fails, 0, fails.size());
 	}
