+++ src/test/java/de/tilman_neumann/jml/factor/hart/HartFast2Mult2Test.java
@@ -13,6 +13,7 @@
  */
 package de.tilman_neumann.jml.factor.hart;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.junit.BeforeClass;
@@ -44,7 +45,7 @@ public class HartFast2Mult2Test extends FactorTestBase {
 	
 	@Test
 	public void testSmallestComposites() {
-		List<Integer> fails = testFullFactorizationOfComposites(100000);
+		List<BigInteger> fails = testFullFactorizationOfComposites(100000);
 		assertEquals("Failed to factor n = " + fails, 0, fails.size());
 	}
