+++ src/test/java/de/tilman_neumann/jml/factor/pollardRho/PollardRhoBrentModBlockTest.java
@@ -13,6 +13,7 @@
  */
 package de.tilman_neumann.jml.factor.pollardRho;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.junit.BeforeClass;
@@ -33,7 +34,7 @@ public class PollardRhoBrentModBlockTest extends FactorTestBase {
 	
 	@Test
 	public void testSmallestComposites() {
-		List<Integer> fails = testFullFactorizationOfComposites(100000);
+		List<BigInteger> fails = testFullFactorizationOfComposites(100000);
 		assertEquals("Failed to factor n = " + fails, 0, fails.size());
 	}
