+++ src/test/java/de/tilman_neumann/jml/factor/squfof/SquFoF31Test.java
@@ -15,6 +15,7 @@ package de.tilman_neumann.jml.factor.squfof;
 
 import static org.junit.Assert.assertEquals;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.junit.BeforeClass;
@@ -33,7 +34,7 @@ public class SquFoF31Test extends FactorTestBase {
 	
 	@Test
 	public void testSmallestComposites() {
-		List<Integer> fails = testFullFactorizationOfComposites(100000);
+		List<BigInteger> fails = testFullFactorizationOfComposites(100000);
 		assertEquals("Failed to factor n = " + fails, 0, fails.size());
 	}
